backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
pydantic-settings==2.0.3
python-jose==3.3.0
python-multipart==0.0.6
passlib==1.7.4
bcrypt==4.0.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
asyncpg==0.29.0
alembic==1.12.1
mne==1.5.1
numpy==1.26.1
pandas==2.1.2
scikit-learn==1.3.2
librosa==0.10.1
transformers==4.35.0
tensorboard==2.14.0
influxdb-client==1.36.1
redis==5.0.1
kafka-python==2.0.2
python-dotenv==1.0.0 